django_backend/requirements.txt
# Django Backend Requirements
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==2.4.2
Pillow==12.1.1

# Para desenvolvimento
python-dateutil==2.9.0
